Kernel: Xiaomi kernel changes for Redmi Y3 and Redmi7 Android P.
The Patch based on QualComm release TAG: LA.UM.7.6.2.r1-04200-89xx.0.
The kernel config file used are onc-perf_defconfig and onclite-perf_defconfig.

+++ include/linux/mm_types.h
@@ -410,7 +410,7 @@ struct mm_struct {
 #ifdef CONFIG_SPECULATIVE_PAGE_FAULT
 	rwlock_t mm_rb_lock;
 #endif
-	u32 vmacache_seqnum;                   /* per-thread vmacache */
+	u64 vmacache_seqnum;                   /* per-thread vmacache */
 #ifdef CONFIG_MMU
 	unsigned long (*get_unmapped_area) (struct file *filp,
 				unsigned long addr, unsigned long len,
